striptcc: checksumming in filecheck() requires read+write access.
With write-only access, the checksumming process would always fail causing incorrect checkSumAdjustment being written to the .ttf files produced. It would always be the base 0xB1B0AFBA value.

In practice, I'm not sure how big of an issue this really is, as macOS seems perfectly happy to use the fonts as is. Examination of several stock macOS .ttf files shows that even some of them have technically incorrect values, so who knows...